backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
google-generativeai==0.5.0
requests==2.32.3
beautifulsoup4==4.13.3
PyPDF2==3.0.1